provide a cleaner approach from scratch within the whole project
