+++ src/bench.c
@@ -1,8 +1,12 @@
+//
+// Created by erena on 13.09.2024.
+//
 
 #include "bench.h"
 
+
 // Benchmarks from Alexandria
-static char* benchmarkfens[52] = {
+char* benchmarkfens[52] = {
         "r3k2r/2pb1ppp/2pp1q2/p7/1nP1B3/1P2P3/P2N1PPP/R2QK2R w KQkq a6 0 14",
         "4rrk1/2p1b1p1/p1p3q1/4p3/2P2n1p/1P1NR2P/PB3PP1/3R1QK1 b - - 2 24",
         "r3qbrk/6p1/2b2pPp/p3pP1Q/PpPpP2P/3P1B2/2PB3K/R5R1 w - - 16 42",
@@ -57,7 +61,7 @@ static char* benchmarkfens[52] = {
         "7k/8/7P/5B2/5K2/8/8/8 b - - 0 175"
 };
 
-static inline void benchmark(int depth, board* position, time* time) {
+void benchmark(int depth, board* position, time* time) {
     U64 totalNodes = 0;
 
     int benchStartTime = getTimeMiliSecond();
@@ -75,7 +79,7 @@ static inline void benchmark(int depth, board* position, time* time) {
 
         //clear counter moves
         clearCounterMoves();
-        totalNodes += nodes;
+        totalNodes += searchNodes;
         //printf("position: %s\n", benchmarkfens[i]);
     }
     int benchFinishTime = getTimeMiliSecond() - benchStartTime;
