+++ src/bit_manipluation.c
@@ -1,33 +1,20 @@
 //
-// Created by erena on 20.04.2024.
+// Created by erena on 13.09.2024.
 //
-#pragma once
 
-#include <stdio.h>
-
-
-#define max(x, y) ((x) > (y) ? (x) : (y))
-#define min(x, y) ((x) < (y) ? (x) : (y))
-#define myAbs(x) ((x) < 0 ? -(x) : (x))
+#include "bit_manipluation.h"
 
 
-#define U64 unsigned long long
-
-#define popBit(bitboard, square) ((bitboard) &= ~(1ULL << (square)))
-#define setBit(bitboard, square) bitboard |= (1ULL << square)
-#define getBit(bitboard, square) bitboard & (1ULL << square)
-
-
-inline int countBits(U64 bitboard) {
+int countBits(U64 bitboard) {
     return __builtin_popcountll(bitboard);
 }
 
 // get least significant 1st bit index
-inline int getLS1BIndex(U64 bitboard) {
+int getLS1BIndex(U64 bitboard) {
     return __builtin_ctzll(bitboard);
 }
 
-static inline void printBitboard(U64 bitboard) {
+void printBitboard(U64 bitboard) {
     printf("\n");
     for (int y = 0, coordinate = 8; y < 8; y++, coordinate--) {
         printf("%d  ", coordinate);
@@ -41,7 +28,7 @@ static inline void printBitboard(U64 bitboard) {
     printf("\n   Bitboard: %llu decimal\n\n", bitboard);
 }
 
-inline U64 setOccupancy(int index, int bitsInMask, U64 attackMask) {
+U64 setOccupancy(int index, int bitsInMask, U64 attackMask) {
     U64 occupancy = 0ULL;
     for (int count = 0; count < bitsInMask; count++) {
         int square = getLS1BIndex(attackMask);
