--- src/bit_manipulation.c
@@ -1,5 +1,0 @@
-//
-// Created by erena on 29.05.2024.
-//
-#include "bit_manipulation.h"
-
