--- src/see.c
@@ -1,10 +1,0 @@
-//
-// Created by erena on 31.05.2024.
-//
-
-
-#include "see.h"
-
-
-
-
