--- src/test/see_test.h
@@ -1,12 +1,0 @@
-//
-// Created by erena on 31.05.2024.
-//
-#pragma once
-
-typedef struct {
-    char *fen;
-    int move;
-    int score;
-} see_test_case;
-
-int test_see();
