--- src/values.c
@@ -1,5 +1,0 @@
-//
-// Created by erena on 31.05.2024.
-//
-
-#include "values.h"
