--- src/zobrist.c
@@ -1,10 +1,0 @@
-//
-// Created by erena on 18.07.2024.
-//
-
-#include "zobrist.h"
-
-
-
-
-
